Rename file ble.h into nrf_ble.h.
This change will mitigate issues when nrf51-sdk is compiled on mbed-classic.
