+++ source/nordic_sdk/components/ble/ble_services/ble_dfu/ble_dfu.h
@@ -50,7 +50,7 @@
 #include <stdint.h>
 #include "ble_gatts.h"
 #include "ble_gap.h"
-#include "ble.h"
+#include "nrf_ble.h"
 #include "ble_srv_common.h"
 
 #define BLE_DFU_SERVICE_UUID                 0x1530                       /**< The UUID of the DFU Service. */
