+++ source/nordic_sdk/components/ble/common/ble_conn_state.c
@@ -34,7 +34,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <string.h>
-#include "ble.h"
+#include "nrf_ble.h"
 #include "sdk_mapped_flags.h"
 #include "app_error.h"
