+++ source/nordic_sdk/components/ble/common/ble_conn_state.h
@@ -62,7 +62,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-#include "ble.h"
+#include "nrf_ble.h"
 #include "sdk_mapped_flags.h"
 
 /**@brief Connection handle statuses.
