+++ source/nordic_sdk/components/ble/common/ble_sensor_location.h
@@ -31,7 +31,7 @@
  */
 
  /* Attention!
-*  To maintain compliance with Nordic Semiconductor ASA’s Bluetooth profile
+*  To maintain compliance with Nordic Semiconductor ASA's Bluetooth profile
 *  qualification listings, this section of source code must not be modified.
 */
