+++ source/nordic_sdk/components/ble/common/ble_srv_common.c
@@ -31,7 +31,7 @@
  */
 
 /* Attention!
-*  To maintain compliance with Nordic Semiconductor ASA’s Bluetooth profile
+*  To maintain compliance with Nordic Semiconductor ASA's Bluetooth profile
 *  qualification listings, this section of source code must not be modified.
 */
 
@@ -39,7 +39,7 @@
 #include <string.h>
 #include "nordic_common.h"
 #include "app_error.h"
-#include "ble.h"
+#include "nrf_ble.h"
 
 uint8_t ble_srv_report_ref_encode(uint8_t                    * p_encoded_buffer,
                                   const ble_srv_report_ref_t * p_report_ref)
