+++ source/nordic_sdk/components/ble/device_manager/device_manager.h
@@ -88,7 +88,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include "sdk_common.h"
-#include "ble.h"
+#include "nrf_ble.h"
 #include "ble_gap.h"
 #include "device_manager_cnfg.h"
