+++ source/nordic_sdk/components/ble/peer_manager/id_manager.h
@@ -35,7 +35,7 @@
 
 #include "stdint.h"
 #include "sdk_errors.h"
-#include "ble.h"
+#include "nrf_ble.h"
 #include "ble_gap.h"
 #include "peer_manager_types.h"
