+++ source/nordic_sdk/components/softdevice/common/softdevice_handler/ble_stack_handler_types.h
@@ -46,7 +46,7 @@
 #ifdef BLE_STACK_SUPPORT_REQD
 
 #include <stdlib.h>
-#include "ble.h"
+#include "nrf_ble.h"
 #include "nrf_sdm.h"
 #include "app_error.h"
 #include "app_util.h"
