+++ source/nordic_sdk/components/softdevice/common/softdevice_handler/softdevice_handler.c
@@ -44,7 +44,7 @@
 #elif defined(ANT_STACK_SUPPORT_REQD) 
     #include "ant_interface.h"
 #elif defined(BLE_STACK_SUPPORT_REQD)
-    #include "ble.h"
+    #include "nrf_ble.h"
 #endif
 
 #ifdef NRF51
